docker_requirements.txt
flask==3.0.0
gunicorn==21.2.0
livekit-api==0.17.6
pyjwt==2.9.0
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.5